+++ src/HomeControl.h
@@ -27,12 +27,13 @@
 #include "DeviceDistance.h"
 #include "DeviceAnalogInput.h"
 #include "DeviceDS18B20.h"
+#include "DevicePWM.h"
 
 #define MAX_DEVICES                     40
 #define INPUT_BUFFER_SIZE               500
 #define SERIAL_INPUT_BUFFER_SIZE        50
 #define CONNECTION_TIMEOUT              20*1000L
-#define VERSION                         6
+#define VERSION                         7
 
 #define EEPROM_INITIALIZED_VALUE        255
 #define EEPROM_CONFIG_SET_OFFSET        0